fix: Use quote instead of encode for job search
For job search, use quote instead of encode. Now users can filter LinkedIn searches—for example, using -senior will remove all jobs that contain "senior" in the description or title. This didn't work before because when performing a search, encode replaced spaces with "+", which is a special character in LinkedIn search.